# .well-known/openapi.yaml
openapi: 3.0.1
info:
  title: Bear Notes Plugin API
  description: API for querying local Bear notes
  version: "1.0.0"

servers:
  - url: https://c29ba4fb2eaf.ngrok-free.app

paths:
  /search_notes:
    post:
      summary: 语义检索笔记
      operationId: searchNotes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                args:
                  type: object
                  properties:
                    query:
                      type: string
                      description: 查询关键词
      responses:
        "200":
          description: 成功返回匹配笔记
          content:
            application/json:
              schema:
                type: object
                properties:
                  notes:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        title:
                          type: string
                        content:
                          type: string

  /retrieve_for_rag:
    post:
      summary: 检索上下文（RAG 模式）
      operationId: retrieveForRag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                args:
                  type: object
                  properties:
                    query:
                      type: string
                      description: 查询文本
      responses:
        "200":
          description: 返回相似笔记上下文
          content:
            application/json:
              schema:
                type: object
                properties:
                  context:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        title:
                          type: string
                        content:
                          type: string
                        score:
                          type: number

  /daily_insight_context:
    post:
      summary: 每日洞察生成与语义延展
      operationId: dailyInsightContext
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                args:
                  type: object
                  properties:
                    date:
                      type: string
                      description: 指定日期（可选）
      responses:
        "200":
          description: 返回每日洞察分析
          content:
            application/json:
              schema:
                type: object
                properties:
                  system_prompt:
                    type: string
                  today:
                    type: array
                    items:
                      type: object
                  related:
                    type: array
                    items:
                      type: object

  /agent/session/start:
    post:
      summary: 启动 Boss-Agent 会话
      operationId: startAgentSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - goal
              properties:
                goal:
                  type: string
                  description: 老板目标，例如“查找包含AI标签的笔记并做摘要”
      responses:
        "200":
          description: 返回会话 ID 和首轮汇报
          content:
            application/json:
              schema:
                type: object
                properties:
                  sessionId:
                    type: string
                  done:
                    type: boolean
                  status:
                    type: string
                  round:
                    type: integer
                  maxRounds:
                    type: integer
                  report:
                    type: string
                  output:
                    type: string
                  validation:
                    type: object
                    properties:
                      ok:
                        type: boolean
                      reason:
                        type: string

  /agent/session/message:
    post:
      summary: 向 Boss-Agent 会话发送消息
      operationId: messageAgentSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - sessionId
              properties:
                sessionId:
                  type: string
                  description: 会话 ID
                message:
                  type: string
                  description: 指令内容，支持“继续 / 调整... / 结束”
      responses:
        "200":
          description: 返回本轮汇报和会话状态
          content:
            application/json:
              schema:
                type: object
                properties:
                  done:
                    type: boolean
                  status:
                    type: string
                  round:
                    type: integer
                  report:
                    type: string
                  output:
                    type: string
                  validation:
                    type: object
                    properties:
                      ok:
                        type: boolean
                      reason:
                        type: string

  /agent/session/{id}:
    get:
      summary: 获取 Boss-Agent 会话状态
      operationId: getAgentSession
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: 会话 ID
      responses:
        "200":
          description: 返回完整会话状态
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  goal:
                    type: string
                  status:
                    type: string
                  round:
                    type: integer
                  maxRounds:
                    type: integer
                  currentTag:
                    type: string
                  output:
                    type: string
                  lastReport:
                    type: string
                  chatHistory:
                    type: array
                    items:
                      type: object